% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cleanschedule_met}
\alias{get_cleanschedule_met}
\title{Wrangling Function}
\usage{
get_cleanschedule_met(sport_schedule, met_values, mapping)
}
\value{
A data frame containing the activities per day with their correspondingly Met values.
}
\description{
A function that cleans the output (data rame) extracted from the webscrape_sports function. 
This function maps the MET values into the data in order to calculate the calorie burn amount per activity.
}
\examples{
# Use the data frame retrieved from the webscrape_sports function
get_cleanschedule_met()
}
