% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_values.R
\docType{data}
\name{met_values}
\alias{met_values}
\title{Metabolic equivalent of task (MET)}
\format{
A data frame with 821 observations and 3 variables:
\describe{
  \item{Activity}{Activity}
  \item{Specific Motion}{Sub-activity/Specific motion} 
  \item{METs}{MET value for each sub-activity}
}
}
\source{
\url{https://golf.procon.org/met-values-for-800-activities/}
}
\usage{
met_values
}
\description{
This table contains the metabolic equivalent of task (MET) of each exercise activity (821 sub-activities in total)
}
\keyword{datasets}
