% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optimize_schedule}
\alias{optimize_schedule}
\title{Optimization Function}
\usage{
optimize_schedule(
  clean_sport_schedule,
  date,
  activity,
  time,
  calburn,
  weight,
  flag_no_duplicate_activities = 0
)
}
\arguments{
\item{date}{}

\item{cleanschedule}{The data frame output from the get_cleanschedule_met. The output of this function should not be modified, so this function can apply the integer optimization technique properly. .}
}
\description{
Function that evaluates the calorie burn per activity and time according to the parameters entered.
}
\examples{
# Define Inputs
calburn <- 500
date <- c('2022-12-14')
activity <- c('Aquagym', 'Zumba', 'Pilates', 'Agrès',
             'Tai ji quan / Tous niveaux',
             'Musculation connectée / 1. Introduction',
             'Cirque', 'Aviron / Débutants', 'Salsa cubaine / Débutants')
weight <- 50
time <- c('07:00 – 08:00', '08:00 – 09:00', '12:00 – 13:00', '13:00 – 14:00',
         '17:00 – 18:00', '18:00 – 19:00', '19:00 – 20:00')
flag_no_duplicate_activities <- 1
load(here::here("data/clean_sport_schedule.rda"))
optimize_output <- optimize_schedule(clean_sport_schedule, date, activity, time, calburn, weight,flag_duplicate_activities)
}
