% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport_schedule.R
\docType{data}
\name{sport_schedule}
\alias{sport_schedule}
\title{Data of activity schedule in the UNIL sport center}
\format{
A data frame with 4 variables:
\describe{
  \item{Date}{Date of each activity}
  \item{Timetable}{The time at which activity begins and ends} 
  \item{Activity}{Activity name}
  \item{Location}{Location of each activity}
}
}
\source{
\url{https://sport.unil.ch/}
}
\usage{
sport_schedule
}
\description{
This table contains details of the sport activities in the UNIL sport center over 7 days
}
\keyword{datasets}
