% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{startApp}
\alias{startApp}
\title{Start Function for Shiny Application}
\usage{
startApp()
}
\value{
A dashboard for the user to interact with
}
\description{
This is a function that automatically runs the shiny function created and opens the dashboard for the user
}
\author{
Arturo Garcia Luna Beltran Manunpat Sirijaturaporn Nadine Obeid Katti Irastorza
}
