% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{UnilSports_gui}
\alias{UnilSports_gui}
\title{Shiny Application function for UnilSports package}
\usage{
UnilSports_gui(clean_sport_schedule)
}
\arguments{
\item{clean_sport_schedule}{The data frame output from the get_cleanschedule_met. The output of this function should not be modified, so this function can apply the integer optimization technique properly.}
}
\value{
a shiny web application dashboard that provides the user with the available sports classes based on certain parameters entered by the user.
}
\description{
This is a function that provides a dashboard for users to input their desired time tables, classes, calories burned to get a set of classes that adhere to the respective requirements
}
\examples{
clean_sport_schedule <- get_cleanschedule_met(sport_schedule,met_values)
}
\author{
Arturo Garcia Luna Beltran Manunpat Sirijaturaporn Nadine Obeid Katti Irastorza
}
