% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{webscrape_MET}
\alias{webscrape_MET}
\title{Webscraping function for MET Values}
\usage{
webscrape_MET()
}
\value{
A data frame containing the MET values per sport activity.
}
\description{
A function that returns the MET values table of more than 800 activities webscrapped from the golf.procon.org website.
}
\examples{
Run the function as it is without any parameter
webscrape_MET()
}
