% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_server}
\alias{build_server}
\title{Server function for UnilSports Shiny Application}
\usage{
build_server(clean_sport_schedule)
}
\arguments{
\item{clean_sport_schedule}{The data frame output from the get_cleanschedule_met. The output of this function should not be modified, so this function can apply the integer optimization technique properly.}
}
\value{
a shiny server output
}
\description{
This is a function that builds a server for the UnilSports shiny application. This function is then used as the secoond parameter in the  UnilSports_gui function.
}
\examples{
build_server(clean_sport_schedule)
}
