% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cleanschedule_met}
\alias{get_cleanschedule_met}
\title{Wrangling Function}
\usage{
get_cleanschedule_met()
}
\value{
A full schedule of the sports provided by Unil
}
\description{
A function that cleans the data scraped from the webpage of Unil sports and produces a proper dataset mapped to another dataset web-scraped for met values. This function does not need any objects to be run, it will run automatically when called.
}
\examples{
get_cleanschedule_met()
}
