% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cleanschedule_met}
\alias{get_cleanschedule_met}
\title{Wrangling Function}
\usage{
get_cleanschedule_met(sport_schedule, met_values)
}
\arguments{
\item{sport_schedule}{Data frame containing the sports schedule of the number of days selected obtained when running webscrape_sports function}

\item{met_values}{Data frame containing the MET values per sport activity obtained when running webscrape_MET function.}
}
\value{
A data frame containing the activities per day with their correspondingly Met values.
}
\description{
A function that cleans the output (data rame) extracted from the webscrape_sports function. 
    This function maps the MET values into the data in order to calculate the calorie burn amount per activity.
}
\examples{
get_cleanschedule_met()
}
\author{
Arturo Garcia Luna Beltran Manunpat Sirijaturaporn Nadine Obeid Katti Irastorza
}
