% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{UnilSports_gui}
\alias{UnilSports_gui}
\title{Shiny Application function for UnilSports package}
\usage{
UnilSports_gui(sport_schedule, met_values, clean_sport_schedule)
}
\arguments{
\item{sport_schedule}{}

\item{met_values}{}

\item{clean_sport_schedule}{}
}
\value{
a shiny web application dashboard that provides the user with the available sports classes based on ceertain parameters entered by the user.
}
\description{
This is a function that provides a dashboard for users to input their desired time tables, classes, calories burned to get a set of classes that adhere to the respective requirements
}
\examples{
Run the below functions from the UnilSports package
sport_schedule <- webscrape_sports()
met_values     <- webscrape_MET()
clean_sport_schedule <- get_cleanschedule_met(sport_schedule,met_values)
UnilSports_gui(sport_schedule, met_values, clean_sport_schedule)
}
