% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{webscrape_sports}
\alias{webscrape_sports}
\title{Web-scraping function for Sports Schedule}
\usage{
webscrape_sports(days = 14, ...)
}
\arguments{
\item{days}{The number of days that you want to retrieve from the Unil Sports Center website}

\item{...}{A parameter that allows the user to enter additional arguments that are not defined in the function}
}
\value{
A data frame containing the sports schedule of the number of days selected
}
\description{
A function that returns the sports table schedule from the Unil Sports Center
}
\examples{
webscrape_sports(days= 14)
}
