% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{webscrape_sports}
\alias{webscrape_sports}
\title{Webscraping function for Sports Schedule}
\usage{
webscrape_sports(days = 7)
}
\arguments{
\item{days}{The number of days that you want to retrieve from the Unil Sports Center website. The defaut number of days is 7.}
}
\value{
A data frame containing the sports schedule of the number of days selected
}
\description{
A function that returns the sports table schedule from the Unil Sports Center.
}
\examples{
webscrap_sports()
}
